LOAN MANAGEMENT SYSTEM UPDATED

+++ src/main/java/uz/pdp/loan_management_system/repository/AuthUserRepository.java
@@ -5,6 +5,6 @@ import uz.pdp.loan_management_system.entity.AuthUser;
 
 import java.util.Optional;
 
-public interface AuthUserRepository extends JpaRepository<AuthUser, Integer> {
+public interface AuthUserRepository extends JpaRepository<AuthUser, Long> {
     Optional<AuthUser> findByUsername(String username);
 }
